Reorganized project, added some loaders, and some new utility functions.
Fixed a minor CSS oddity with the miniloader background in admin.css.

Added a CSSLoader class for assisting in loading CSS from a file, as well as enqueue them automatically for pages.

Added a JSLoader, to load JS files, as well as enqueing them automatically for pages.

Added an IntervalType enum class, to handle setting what kind of interval should be used for the plugin file integrity check.

Added support for loading webcomponenets via the Polymer JS library.

Added more robust page detection funcitons, as WordPress can be complicated when detecting what the page is on the user / admin side.

Fixed an issue where the admin test AJAX page was NOT a static method, and should have been.

Added namespacing for plugin options, as well as enqueued scripts/styles.

Improved some namespacing of the JS code, as well as added TypeScript usage.