Updated plugin to v2.0 and added comments in PHP and JS files
This is a full release, and is comprehensive.

It includes a new architecture, more functions, and full examples on how to use this plugin framework with the TestPlugin, as well as comments throughout for functions/logic that i believe would be a bit confusing.

The main repo description Readme.md will be updated to reflect this big change.